+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `SamplingParams.json_sampler_names`, `SamplingParams.append_to_sampler_names`, `SamplingParams.remove_from_sampler_names`, `SamplingParams.extend_sampler_names`, `SamplingParams.append_to_json_sampler_names`, `SamplingParams.remove_from_json_sampler_names`, and `SamplingParams.extend_json_sampler_names`.
 - Added `jsonSampler` and `jsonSamplerMap` to `libcasm.monte.sampling`.
 - Added selected event data sampling methods to `libcasm.monte.sampling`.
+- Added memory usage to results.
 
 ### Changed